Revert "block: anxiety-iosched: don't use size_t type for starvation"
This reverts commit c778e2f578c0db2cb711b888d2bf59090c82ea8e.

+++ block/anxiety-iosched.c
@@ -13,7 +13,7 @@
 
 struct anxiety_data {
 	struct list_head queue[2];
-	unsigned int writes_starved;
+	size_t writes_starved;
 };
 
 static void anxiety_merged_requests(struct request_queue *q, struct request *rq, struct request *next)
